+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 
 - `positionsItselfAndKeepDistance` little improvements
+- `bonfireCamera.shake` improvements
 
 # 3.0.0-beta.10
 - `Follower` mixin improvements.